refactor: Move more blog components to blog directory.
