+++ components/blog/CopyableHeader/CopyableHeader.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/breakpoints.module.scss';
+@import '../../../styles/breakpoints.module.scss';
 
 .h2 {
     margin-top: var(--spacing-7);
