+++ components/blog/OpenToWork/OpenToWork.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/breakpoints.module.scss';
+@import '../../../styles/breakpoints.module.scss';
 
 .wrapper {
     background: var(--otw-background-color);
